requirements.txt
langchain
langchain-openai
langchain-core
langgraph
pydantic>=2.0
python-dotenv
httpx
langsmith
# 如果这台机器也负责跑本地模型，再取消下面这行的注释
# vllm